+++ sma-em-dev/run.py
@@ -7,6 +7,7 @@ import struct
 import sys
 
 import sensors
+from options import OPT, init_options
 from speedwiredecoder import decode_speedwire
 
 MCAST_GRP = "239.12.255.254"
@@ -17,19 +18,27 @@ _LOGGER = logging.getLogger(__name__)
 
 
 class MulticastServerProtocol(asyncio.DatagramProtocol):
+    """Multicast receiver."""
+
     def connection_made(self, _transport):
-        pass
+        """Protocol connected."""
+        _LOGGER.info(
+            "Listening for multicast frames. "
+            "Sensor discovery will be triggered by the first frame."
+        )
 
     def datagram_received(self, data, _addr):
+        """Process frame."""
         try:
             emparts = decode_speedwire(data)
-        except Exception as err:
+        except Exception as err:  # pylint: disable=broad-except
             _LOGGER.warning("Could not decode Speedwire %s", err)
         else:
             asyncio.get_running_loop().create_task(sensors.process_emparts(emparts))
 
 
 def connect_socket():
+    """Connect."""
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM, socket.IPPROTO_UDP)
     sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
     sock.bind(("", MCAST_PORT))
@@ -39,32 +48,37 @@ def connect_socket():
             "4s4s", socket.inet_aton(MCAST_GRP), socket.inet_aton(IPBIND)
         )
         sock.setsockopt(socket.IPPROTO_IP, socket.IP_ADD_MEMBERSHIP, mreq)
-    except BaseException:
+    except BaseException:  # pylint: disable=broad-except
         print("could not connect to mulicast group or bind to given interface")
         sys.exit(1)
     return sock
 
 
 async def main():
-    sensors.startup()
-    logging.basicConfig(
-        level=logging.DEBUG if sensors.OPTIONS["DEBUG"] != 0 else logging.INFO
+    """Addon entry."""
+    init_options()
+
+    sensors.MQ_CLIENT.availability_topic = (
+        f"{sensors.SMA_EM_TOPIC}/{''.join(OPT.sma_serials)}/available"
     )
 
-    LOOP = asyncio.get_event_loop()
-    LOOP.set_debug(True)
+    await sensors.MQ_CLIENT.connect(options=OPT)
 
-    RECONNECT_INTERVAL = sensors.OPTIONS.get(sensors.RECONNECT_INTERVAL, 86400)
+    if OPT.debug == 0:
+        sensors.ic.disable()
+
+    loop = asyncio.get_event_loop()
+    loop.set_debug(True)
 
     while True:
         _sock = connect_socket()
-        _transport, _protocol = await LOOP.create_datagram_endpoint(
+        _transport, _protocol = await loop.create_datagram_endpoint(
             MulticastServerProtocol,
             sock=_sock,
         )
 
         try:
-            await asyncio.sleep(RECONNECT_INTERVAL)
+            await asyncio.sleep(OPT.reconnect_interval)
         finally:
             _transport.close()
             _sock.close()