+++ sma-em-dev/sensors.py
@@ -2,15 +2,14 @@
 import asyncio
 import logging
 import statistics
-import sys
 import time
-from json import dumps, loads
-from pathlib import Path
-from typing import Any, Dict, Sequence
+from typing import Union
 
 import attr
 from icecream import ic
-from mqtt import MQTTClient
+from mqtt_entity import Device, MQTTClient, SensorEntity
+from mqtt_entity.helpers import hass_device_class
+from options import OPT
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -19,17 +18,21 @@ _LOGGER = logging.getLogger(__name__)
 class SWSensor:
     """A speedwire sensor."""
 
-    name = attr.field()
-    mod = attr.field()  # 0 = mean, 1 = max, -1 = min
+    # pylint: disable=too-many-instance-attributes,too-few-public-methods
+
+    name: str = attr.field()
+    mod: str = attr.field()  # 0 = mean, 1 = max, -1 = min
     last_update: int = attr.field(default=0)
-    interval = attr.field(default=5)
-    value = attr.field(default=0)
-    values = attr.field(factory=list)
-    unit = attr.field(default=None)
-    device_class = attr.field(default=None)
+    interval: int = attr.field(default=5)
+    value: Union[int, float] = attr.field(default=0)
+    values: list[Union[int, float]] = attr.field(factory=list)
+    unit: str = attr.field(default="")
+    device_class: str = attr.field(default="")
+    mq_entity: SensorEntity = attr.field(default=None)
 
     @property
-    def id(self):
+    def id(self) -> str:  # pylint: disable=invalid-name
+        """Return the ID."""
         if self.mod == "":
             return self.name
         if self.mod == "avg":
@@ -37,73 +40,73 @@ class SWSensor:
         return f"{self.name}_{self.mod}"
 
 
-SENSORS: Dict[str, Sequence[SWSensor]] = {}
-OPTIONS: Dict[str, Any] = {}
-_MQTT = MQTTClient()
+SENSORS: dict[str, list[SWSensor]] = {}
+MQ_CLIENT = MQTTClient()
 
-
-MQTT_HOST = "MQTT_HOST"
-MQTT_PORT = "MQTT_PORT"
-MQTT_USERNAME = "MQTT_USERNAME"
-MQTT_PASSWORD = "MQTT_PASSWORD"
-SERIALS = "SMA_SERIALS"
-FIELDS = "FIELDS"
-THRESHOLD = "THRESHOLD"
-DEBUG = "DEBUG"
-RECONNECT_INTERVAL = "RECONNECT_INTERVAL"
 
 SMA_EM_TOPIC = "SMA-EM/status"
 
 
-async def mqtt_publish(topic: str, payload: Any, retain: bool = False):
-    await _MQTT.connect(
-        username=OPTIONS[MQTT_USERNAME],
-        password=OPTIONS[MQTT_PASSWORD],
-        host=OPTIONS[MQTT_HOST],
-        port=OPTIONS[MQTT_PORT],
-    )
-    if not topic:
+async def process_emparts(emparts: dict) -> None:
+    """Process emparts from the speedwire decoder."""
+    # pylint: disable=too-many-branches
+    if emparts["protocol"] not in [0x6069, 0x6081]:
+        _LOGGER.info("Ignore protocol %s", hex(emparts["protocol"]))
         return
-    ic(topic, payload)
-    await _MQTT.publish(topic=topic, payload=payload, retain=retain)
 
-
-async def process_emparts(emparts: dict):
-    if (emparts["protocol"] not in [0x6069, 0x6081]):
-        _LOGGER.info(
-            "Ignore protocol %s",
-            hex(emparts["protocol"])
-        )
-        return
     serial = emparts["serial"]
     if not SENSORS.get(serial):
-        SENSORS[serial] = get_sensors(definition=OPTIONS[FIELDS], emparts=emparts)
+        # Discover the sensors
         _LOGGER.info(
+            "Multicast frame received for SMA serial %s, discovered sensors:", serial
+        )
+        SENSORS[serial] = get_sensors(definition=OPT.fields, emparts=emparts)
+        _LOGGER.debug(
             "Discover %s/%s sensors on SMA %s",
             len(SENSORS[serial]),
-            len(OPTIONS[FIELDS]),
+            len(OPT.fields),
             serial,
         )
+        mq_dev = Device(
+            identifiers=[serial, f"sma_em_{serial}"],
+            name="SMA Energy Meter",
+            model="Energy Meter",
+            manufacturer="SMA",
+        )
+
         for sen in SENSORS[serial]:
             _LOGGER.info(" - %s every %ss", sen.id, sen.interval)
-            await hass_discover_sensor(sma_id=serial, sensor=sen)
+            sen.mq_entity = SensorEntity(
+                name=sen.id,
+                device_class=sen.device_class,
+                state_topic=f"{SMA_EM_TOPIC}/{serial}/{sen.id}",
+                unique_id=f"{serial}_{sen.id}",
+                unit_of_measurement=sen.unit,
+                device=mq_dev,
+            )
+
+        await MQ_CLIENT.publish_discovery_info(
+            [s.mq_entity for sl in SENSORS.values() for s in sl]
+        )
         await asyncio.sleep(5)
 
-    push_later = []
+    push_later: list[tuple[SWSensor, Union[int, float], int]] = []
     now = int(time.time())
     for sen in SENSORS[serial]:
         val = emparts.get(sen.name)
+        if val is None:
+            continue
 
         publish = now >= sen.last_update + sen.interval
 
         # check threshold crossing
         smart_s = sen.mod == "" and sen.values
-        if smart_s and val > sen.value + OPTIONS[THRESHOLD]:
+        if smart_s and val > sen.value + OPT.threshold:
             publish = True
-            push_later.append((sen, val, OPTIONS[THRESHOLD]))
-        elif smart_s and val < sen.value - 2 * OPTIONS[THRESHOLD]:
+            push_later.append((sen, val, OPT.threshold))
+        elif smart_s and val < sen.value - 2 * OPT.threshold:
             publish = True
-            push_later.append((sen, val, -2 * OPTIONS[THRESHOLD]))
+            push_later.append((sen, val, -2 * OPT.threshold))
         else:
             sen.values.append(val)
 
@@ -120,7 +123,7 @@ async def process_emparts(emparts: dict):
             sen.value = round(statistics.mean(sen.values), 1)
         sen.values = []
 
-        await mqtt_publish(f"{SMA_EM_TOPIC}/{emparts['serial']}/{sen.id}", sen.value)
+        await MQ_CLIENT.publish(sen.mq_entity.state_topic, sen.value)
 
     if not push_later:
         return
@@ -129,48 +132,12 @@ async def process_emparts(emparts: dict):
     for sen, val, delta in push_later:
         sen.value = round(val, 1)
         ic(sen.name, sen.value, delta)
-        await mqtt_publish(f"{SMA_EM_TOPIC}/{emparts['serial']}/{sen.id}", sen.value)
+        await MQ_CLIENT.publish(sen.mq_entity.state_topic, sen.value)
 
 
-# device_calss:current/energy/power
-# unit_of_measurement
-def hass_device_class(*, unit: str):
-    return {
-        "W": "power",
-        "kW": "power",
-        "kVA": "power",
-        "V": "voltage",
-        "Hz": None,
-    }.get(
-        unit, "energy"
-    )  # kwh, kVa,
-
-
-async def hass_discover_sensor(*, sma_id: str, sensor: SWSensor):
-    topic = f"homeassistant/sensor/{sma_id}/{sensor.id}/config"
-    payload = {
-        "name": sensor.id,
-        "dev_cla": sensor.device_class,
-        "stat_t": f"{SMA_EM_TOPIC}/{sma_id}/{sensor.id}",
-        "unit_of_meas": sensor.unit,
-        "uniq_id": f"{sma_id}_{sensor.id}",
-        # "state_class": "measurement",
-        # "last_reset": "2021-07-30T00:00:00+00:00",
-        "dev": {
-            "ids": [f"sma_em_{sma_id}"],
-            "name": "SMA Energy Meter",
-            "mdl": "Energy Meter",
-            "mf": "SMA",
-        },
-    }
-    if sensor.device_class == "energy":
-        payload["state_class"] = "total_increasing"
-
-    await mqtt_publish(topic, dumps(payload), retain=True)
-
-
-def get_sensors(*, definition: str, emparts: dict):
-    res = []
+def get_sensors(*, definition: list[str], emparts: dict) -> list[SWSensor]:
+    """Create a list of all SWSensors from the definitions and emparts."""
+    res: list[SWSensor] = []
     for sensor_def in definition:
         name, _, mod = sensor_def.partition(":")
 
@@ -178,7 +145,7 @@ def get_sensors(*, definition: str, emparts: dict):
             _LOGGER.info("Unknown sensor: %s", name)
             continue
 
-        sen = SWSensor(name=name, mod=mod, unit=emparts.get(f"{name}unit"))
+        sen = SWSensor(name=name, mod=mod, unit=emparts.get(f"{name}unit", ""))
         if sen.mod in ["min", "max"]:
             sen.interval = 60
         else:
@@ -193,38 +160,3 @@ def get_sensors(*, definition: str, emparts: dict):
 
     ic(res)
     return res
-
-
-def startup():
-    """Read the hassos configuration."""
-    ofile = Path("/data/options.json")
-    options = (
-        loads(ofile.read_text())
-        if ofile.exists()
-        else {
-            MQTT_HOST: "192.168.88.128",
-            MQTT_PASSWORD: sys.argv[1],
-            MQTT_PORT: 1883,
-            MQTT_USERNAME: "hass",
-            SERIALS: [],
-            FIELDS: [
-                "pconsume",
-                "pconsumecounter:max",
-                "pconsume:5",
-                "u1:min",
-            ],
-            THRESHOLD: 80,
-            RECONNECT_INTERVAL: 86400,
-            DEBUG: 0,
-        }
-    )
-    for key, val in options.items():
-        OPTIONS[key] = val
-
-    if OPTIONS[DEBUG] == 0:
-
-        def blank(*args):
-            pass
-
-        global ic
-        ic = blank